requirements.txt
streamlit==1.40.0
strands-agents==0.1.4
strands-agents-tools==0.1.0
python-dotenv==1.0.1
Pillow==10.4.0
mcp==0.9.2
requests==2.32.3
boto3==1.35.80